+++ src/resources/token.ts
@@ -1,7 +1,7 @@
 // File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
-import * as Core from '@blockaid/client/core';
 import { APIResource } from '@blockaid/client/resource';
+import * as Core from '@blockaid/client/core';
 import * as TokenAPI from '@blockaid/client/resources/token';
 import * as EvmAPI from '@blockaid/client/resources/evm/evm';
 
@@ -53,9 +53,9 @@ export namespace TokenScanResponse {
 
 export interface TokenScanParams {
   /**
-   * Token address to validate (EVM / Solana)
+   * Token address to validate (EVM / Solana / Stellar)
    */
-  address: string | string;
+  address: string;
 
   /**
    * The chain name